+++ test/integration/authorization.test.js
@@ -3,7 +3,7 @@
 const expect = require('code').expect
 const Lab = require('lab')
 const lab = exports.lab = Lab.script()
-const utils = require('./../utils')
+const utils = require('../utils')
 const server = require('../../lib/server')
 
 lab.experiment('Authorization', () => {
