+++ test/integration/authorization/authorization.test.js
@@ -3,7 +3,7 @@ const lab = exports.lab = Lab.script()
 
 const server = require('../../../lib/server')
 const Factory = require('../../factory')
-const BuildFor = require('./testBuilder')
+const BuildFor = require('./test-builder')
 
 const organizationId = 'WONKA'
 function Policy (Statement) {
