+++ test/integration/authorization/organizations.test.js
@@ -3,9 +3,10 @@ const lab = exports.lab = Lab.script()
 
 const server = require('../../../lib/server')
 const Factory = require('../../factory')
-const BuildFor = require('./testBuilder')
+const BuildFor = require('./test-builder')
 const utils = require('../../utils')
 const { udaru } = utils
+
 
 const organizationId = 'WONKA'
 function Policy (Statement) {
