+++ test/integration/authorizeOps.test.js
