+++ test/integration/userOps.test.js
