+++ test/unit/routes/monitor.test.js
@@ -3,7 +3,7 @@
 const expect = require('code').expect
 const Lab = require('lab')
 const lab = exports.lab = Lab.script()
-const server = require('../../lib/server')
+const server = require('../../../lib/server')
 
 lab.experiment('monitor', () => {
   lab.test('calling ping should return 200 Ok', (done) => {
